add optional support for "file-magic"
There are multiple Python packages for the "libmagic" library. Unfortunately,
some of them use the same Python package name ("magic") and can not be installed
in parallel.

This code adds support for the "file-magic" pypi package as a fallback when
"python-magic" is not installed. Unfortunately `pyproject.toml` does not provide
a way to express a XOR dependency so regular users will always use "python-magic".
However this change enables Linux distributions like Fedora to simply patch
`pyproject.toml`.